Fix a bug with the price formatter. (#661)
The price formatter was using parseInt for the string-ified price value. This
led to a loss of precision. Specifically, if a price was stored in the KG as
$10.99, our formatter would report it as $10.00. This is corrected by using
parseFloat.

TEST=auto

Updated the unit tests for the formatter to ensure the precision is not lost.

+++ static/js/formatters-internal.js
@@ -527,7 +527,7 @@ export { generateCTAFieldTypeLink };
  */
 export function price(fieldValue = {}, locale) {
   const localeForFormatting = locale || _getDocumentLocale() || 'en';
-  const price = fieldValue.value && parseInt(fieldValue.value);
+  const price = fieldValue.value && parseFloat(fieldValue.value);
   const currencyCode = fieldValue.currencyCode && fieldValue.currencyCode.split('-')[0];
   if (!price || isNaN(price) || !currencyCode) {
     console.warn(`No price or currency code in the price fieldValue object: ${fieldValue}`);
